+++ src/st/wrp_psp/e_misc.c
@@ -72,31 +72,7 @@ void func_8018F510(Entity* entity) {
 
 #include "../entity_unkId14.h"
 
-extern s16 unk15_rot[];
-extern s32 unk15_yVel[];
-void EntityUnkId15(Entity* self) {
-    if (!self->step) {
-        self->flags = FLAG_UNK_2000 | FLAG_UNK_04000000 | FLAG_UNK_08000000;
-        self->palette = PAL_OVL(0x195);
-        self->animSet = ANIMSET_DRA(5);
-        self->animCurFrame = 1;
-        self->drawMode = DRAW_TPAGE;
-        self->drawFlags = FLAG_DRAW_ROTX | FLAG_DRAW_ROTY;
-        self->rotX = unk15_rot[self->params];
-        self->rotY = self->rotX;
-        self->velocityY = unk15_yVel[self->params];
-        self->step++;
-    } else {
-        self->posY.val -= self->velocityY;
-        self->animFrameDuration++;
-        if (!(self->animFrameDuration % 2)) {
-            self->animCurFrame++;
-        }
-        if (self->animFrameDuration > 36) {
-            DestroyEntity(self);
-        }
-    }
-}
+#include "../entity_unkId15.h"
 
 extern u16 g_InitializeEntityData0[];
